% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateEVfilesFromTemplate.R
\name{CreateEVFilesFromTemplate}
\alias{CreateEVFilesFromTemplate}
\title{Create EV Files from a Template}
\usage{
CreateEVFilesFromTemplate(
  FileSetName = NULL,
  SvVarName = NULL,
  TsVarName = NULL,
  UpperLineName = NULL,
  LowerLineName = NULL,
  CreateAnalysisRegion = TRUE
)
}
\arguments{
\item{FileSetName}{character vector providing name of fileset on which to
conduct operations.}

\item{SvVarName}{character vector naming the Sv variable to be created and
edited. Must match the name of the variable in the template used.}

\item{TsVarName}{character vector naming the single target variable to be created
and edited. Must match the name of the variable in the template used.}

\item{UpperLineName}{character vector naming the upper line to be used for
exclusion. Also used in creation of line-relative region.}

\item{LowerLineName}{character vector naming the upper line to be used for
exclusion. Also used in creation of line-relative region.}

\item{CreateAnalysisRegion}{logical vector indicating whether or not the
user wants to create a line-relative region between UpperLineName and LowerLineName
and spanning the entire echogram.}
}
\value{

}
\description{
Create EV Files from a Template
}
\details{
This function can be used to create EV files from a template, pick a
bottom line from an Sv variable named by the user, set exclusion lines for that
Sv variable and a TS variable named by the user, and create a line-relative
analysis region covering the entire span of the data in the EV file.

Though not included in the parameter list above, the user will be prompted to
select
1. A pre-existing folder in which to save the EV files.
2. The folder in which raw data files are stored. This folder must have
subfolders with transect names that contain the raw data for the respective
transects. It works well for multi-ship surveys to have a folder called transects
or raw data, a next-level folder for each ship, and next-level folders for each transect.
3. The EV file template
4. The ecs file associated with the template.
}
\author{
David Warner and Kristy Phillips
}
