% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVFreqDistRegionsByCellsExport.R
\name{EVFreqDistRegionsByCellsExport}
\alias{EVFreqDistRegionsByCellsExport}
\title{Function to export TS frequency distribution.}
\usage{
EVFreqDistRegionsByCellsExport(EVFile, TSacoVarName, regionClassName, exportFn)
}
\arguments{
\item{EVFile}{}

\item{TSacoVarName}{}

\item{regionClassName}{}

\item{exportFn}{}
}
\value{
This function exports single target frequency distribution by region-cells.
}
\description{
This function runs region-by cell exports of TS frequency distribution 
for EV files identified in the data directory selected and exports them to 
the output directory selected. Can be called on its own but is also used in 
the function ExportSvTSdistRegionbyCell.
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
