% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVGetCalibrationFileName.R
\name{EVGetCalibrationFileName}
\alias{EVGetCalibrationFileName}
\title{Get Name of ECS File for an EV File}
\usage{
EVGetCalibrationFileName(EVFile, filesetName)
}
\arguments{
\item{EVFile}{Echoview file COM object}
}
\description{
This reads the name and path of the ECS file for a given
EV file.
}
\details{
None to mention
}
