% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/EVIntegrationByRegionsByCellsExportUseEVFileThresholds.R
\name{EVIntegrationByRegionsByCellsExportUseEVFileThresholds}
\alias{EVIntegrationByRegionsByCellsExportUseEVFileThresholds}
\title{Export integration by regions by cells from an Echoview acoustic variable}
\usage{
EVIntegrationByRegionsByCellsExportUseEVFileThresholds(
  EVFile,
  acoVarName,
  regionClassName,
  exportFn
)
}
\arguments{
\item{EVFile}{An Echoview file object}

\item{acoVarName}{A string containing the name of an Echoview acoustic variable}

\item{regionClassName}{A string containing the name of an Echoview region class}

\item{exportFn}{export filename and path}
}
\value{
a list object with one element, $msg: message for processing log.
}
\description{
This function performs integration by regions by cells for a specified region class and exports the results using COM scripting. The difference between this function and the original
function EVIntegrationByRegionsByCellsExport is that this version does not require the
user to provide a data threshold. This function was created because it was not clear
if the original function used the thresholds applied in the EV file or not.
}
\examples{
\dontrun{
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVFile <- EVOpenFile(EVAppObj,'~~/KAOS/KAOStemplate.EV')$EVFile
EVIntegrationByRegionsByCellsExport(EVFile, "120 aggregations", "aggregations", exportFn = "~~/KAOS/EVIntegrationByRegionsExport_example.csv")
}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVOpenFile}}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
