% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportAllSingleTargets.R
\name{ExportAllSingleTargets}
\alias{ExportAllSingleTargets}
\title{Export Properties for All Single Targets in Selected EV Files}
\usage{
ExportAllSingleTargets(
  dat.dir = NULL,
  out.dir = NULL,
  TargetsVarName = NULL,
  start.ping = 1,
  end.ping = 1e+06
)
}
\arguments{
\item{dat.dir}{Character vector representing name of the directory with EV files. This
must be the folder containing the EV file or files, not an EV file itself. Will search
the folder recursively to find any EV files.}

\item{out.dir}{Character vector representing name of the directory to which single
target exports should be saved.}

\item{TargetsVarName}{Character vector representing name of the single targets variable.}

\item{start.ping}{Integer vector stating the starting ping of the analysis domain.}

\item{end.ping}{Integer vector stating the end ping of the analysis domain.}
}
\value{
This function will open an Echoview instance and export the properties of all single targets
from the names single target variable in each EV file.
}
\description{
Export Properties for All Single Targets in Selected EV Files
}
