% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetEVRawPaths.R
\name{GetEVRawPaths}
\alias{GetEVRawPaths}
\title{Get the names of raw data files used in the EV file from
which sphere target data were exported.}
\usage{
GetEVRawPaths(file.list)
}
\arguments{
\item{file.list}{This is essentially one EV file selected by the user.}
}
\value{
Returns the full filename for the raw data files in the EV file
represented by "file.list".

#' @details
This function is a helper function for Sphere2BeamPattern and is not called
on its own.
}
\description{
Get the names of raw data files used in the EV file from
which sphere target data were exported.
}
