% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{inputs}
\alias{inputs}
\title{Beam Pattern Estimates from Calibration Sphere Data}
\usage{
inputs()
}
\description{
Estimate 3 dB beam angles, angle offsets, and TS gain using
the formula employed by Simrad Lobe program. Creates graphs
showing sphere TS across the beam and a summary text file that
includes numerous pieces of information to help make the process
reproducible.
}
\details{
This function is used to prompt the user for input about the data they wish
to process and the source information.
}
