% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateDataPathsAndECS.R
\name{UpdateDataPathAndECS}
\alias{UpdateDataPathAndECS}
\title{Update EV File Data Paths and Change ECS file}
\usage{
UpdateDataPathAndECS(
  file.list,
  filesetName = "Primary Fileset",
  UpdateECS = FALSE
)
}
\arguments{
\item{filesetName}{Character vector for name of fileset}

\item{UpdateECS}{Logical stating whether or not to update the ECS file}
}
\value{
nothing
}
\description{
This function can be used to add a new path for data files to the
EV file properties for a list of EV files.
}
\details{
This function assumes a specific file storage pattern that may not 
match that of many or most users. The EV files and ECS files are stored as 
"C:/Acoustics/2019/vessel", where vessel is the name of the vessel and the 
folder contains a vessel specific ECS file. Some use ECS file specific to
an EV file. Data files are stored, as an example, as "C:/Acoustics/2019/raw_data",
with sub folders in the "raw_data" folder whose names match the names of the EV
files. This approach is set up so that EV files are each a distinct transect,
}
